fold orchestrator into oracle
