+++ oracle/README.md
@@ -8,7 +8,7 @@ The Orchestrator is the central component of the oracle. It is responsible for:
 
 ## Configuration
 
-At a high level the orchestrator is configured with a `oracle.json` file that contains all providers that need to be instantiated. To read more about the configuration of `oracle.json`, please refer to the [oracle configuration documentation](./../config/README.md).
+At a high level the orchestrator is configured with a `oracle.json` file that contains all providers that need to be instantiated. To read more about the configuration of `oracle.json`, please refer to the [oracle configuration documentation](config/README.md).
 
 Each provider is instantiated using the `PriceAPIQueryHandlerFactory`, `PriceWebSocketQueryHandlerFactory`, and `MarketMapFactory` factory functions. Think of these as the constructors for the providers.
